backend/requirements.txt
fastapi[all] 
uvicorn 
sqlalchemy
mysqlclient
pymysql 
mysql-connector-python
redis 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
bcrypt 
python-jose[cryptography] 
python-dotenv